requirements.txt
PyFunctional==1.2.0
